Update README.md
TechNet Gallery is read-only and will be retired soon. Didn't want to lose this bit of information along with the gallery. I tried to maintain the formatting present in the original post.

Apologies if this is not proper use of this MD readme file.

Source content (soon to be gone):
https://gallery.technet.microsoft.com/Migrate-Group-Policy-2b5067d8#content

TechNet Gallery retirement and closure information:
https://docs.microsoft.com/en-us/teamblog/technet-gallery-retirement